bnd.hashes: Always use Long for hash values
They are written as List<Long> to resource indexes and read back in that
way. By being consistent in using Long, we get consistent results for
equals and hashCode on capabilities.

This is important to the resolve context to avoid duplicate resources
being shown to the resolver.

+++ biz.aQute.bndlib/src/aQute/bnd/classindex/ClassIndexerAnalyzer.java
@@ -31,13 +31,13 @@ public class ClassIndexerAnalyzer implements AnalyzerPlugin {
 			if (classspace.isEmpty())
 				continue;
 
-			List<Integer> hashes = new ArrayList<>();
+			List<Long> hashes = new ArrayList<>();
 			for (Clazz c : classspace) {
 				TypeRef tr = c.getClassName();
 				if (tr.isArray() || tr.isNested())
 					continue;
 
-				int hash = tr.getShorterName()
+				long hash = tr.getShorterName()
 					.hashCode();
 				hashes.add(hash);
 			}
